recommitting migrate

+++ db/migrate/20190414203717_add_interest_expense_to_financial_period.rb
@@ -1,5 +1,5 @@
 class AddInterestExpenseToFinancialPeriod < ActiveRecord::Migration[5.2]
   def change
-    add_column :financial_periods, :interest_expense_net, :integer
+    add_column :financial_periods, :interest_expense_net, :bigint
   end
 end
